backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
python-dotenv>=1.2.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=5.0.0
git+https://github.com/celery/django-celery-beat.git@main#egg=django-celery-beat
jsonschema>=4.20.0
croniter>=2.0.0
cryptography>=41.0.0
requests>=2.31.0
supabase>=2.0.0
postgrest>=0.16.0
realtime>=2.0.0
psutil>=5.9.0
slack-sdk>=3.27.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
jinja2>=3.1.0
tenacity>=8.2.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
deepseek>=0.1.0

